refactor build command prerender lifecycle
