+++ packages/cli/test/cases/build.config.prerender/build.config.prerender.spec.js
@@ -10,7 +10,7 @@
  *
  * User Config
  * {
- *   prerender: 'false'
+ *   prerender: false
  * }
  *
  * User Workspace
